Extract shared runtime summarizer helpers
Deduplicate option normalization, middle clipping, secret masking, summary truncation, and streamed text collection across terminal/file/git summarizers while preserving existing fallback behavior.
