+++ packages/agent-sdk-ts/src/sdk/runtime/gitChangeSummarizer.ts
@@ -4,6 +4,13 @@ import { promisify } from 'node:util';
 import type { ChatCompletionRequest, LLMClient } from '../llm';
 import { getGeminiClient } from './geminiClient';
 import type { SecretRegistry } from './SecretRegistry';
+import {
+  clipTextMiddle,
+  collectStreamedText,
+  maskSecrets,
+  resolveNonNegativeIntOption,
+  truncateSummary,
+} from './summarizerCommon';
 
 const execFileAsync = promisify(execFile);
 
@@ -49,46 +56,6 @@ const DEFAULT_MAX_OVERALL_CHARS = 1_200;
 const DEFAULT_MAX_FILE_SUMMARY_CHARS = 400;
 const DEFAULT_MAX_FILES = 15;
 const CLIP_MARKER = '<diff clipped>';
-
-const resolveNonNegativeIntOption = (value: unknown, defaultValue: number): number => {
-  if (value === undefined) return defaultValue;
-  if (typeof value !== 'number' || !Number.isFinite(value)) return defaultValue;
-  return Math.max(0, Math.trunc(value));
-};
-
-const clipTextMiddle = (text: string, maxChars: number): string => {
-  if (maxChars <= 0) return '';
-  if (text.length <= maxChars) return text;
-  const markerBudget = CLIP_MARKER.length + 2;
-  if (maxChars < markerBudget) return text.slice(0, maxChars);
-  const available = maxChars - markerBudget;
-  const headLen = Math.ceil(available / 2);
-  const tailLen = Math.floor(available / 2);
-  const head = text.slice(0, headLen);
-  const tail = tailLen === 0 ? '' : text.slice(-tailLen);
-  return `${head}\n${CLIP_MARKER}\n${tail}`;
-};
-
-const maskSecrets = (text: string, secrets: SecretRegistry): string => {
-  let masked = text;
-  const values = secrets
-    .getRegisteredValues()
-    .map((value) => value.trim())
-    .filter((value) => value.length > 0)
-    .sort((a, b) => b.length - a.length);
-  for (const value of values) {
-    masked = masked.replaceAll(value, '***');
-  }
-  return masked;
-};
-
-const truncateSummary = (text: string, maxChars: number): string => {
-  if (maxChars <= 0) return '';
-  const trimmed = text.trim();
-  if (trimmed.length <= maxChars) return trimmed;
-  if (maxChars === 1) return '…';
-  return trimmed.slice(0, maxChars - 1) + '…';
-};
 
 const defaultExecFileText = async (command: string, args: string[], cwd?: string): Promise<string> => {
   const { stdout } = await execFileAsync(command, args, {
@@ -198,7 +165,7 @@ export async function summarizeGitChangesWithGeminiFlash(
   const patchText = patch.trimEnd();
   if (!nameStatusText && !patchText) return undefined;
 
-  const clippedPatch = clipTextMiddle(patchText, maxDiffChars);
+  const clippedPatch = clipTextMiddle(patchText, maxDiffChars, CLIP_MARKER);
   const rawPrompt = [
     'Summarize a multi-commit git change set for an IDE UI.',
     '',
@@ -219,7 +186,7 @@ export async function summarizeGitChangesWithGeminiFlash(
     clippedPatch || '(empty)',
   ].join('\n');
 
-  const safePrompt = clipTextMiddle(maskSecrets(rawPrompt, options.secrets), maxPromptChars);
+  const safePrompt = clipTextMiddle(maskSecrets(rawPrompt, options.secrets), maxPromptChars, CLIP_MARKER);
   const request: ChatCompletionRequest = {
     systemPrompt: 'You summarize git diffs for an IDE UI.',
     messages: [{ role: 'user', content: [{ type: 'text', text: safePrompt }] }],
@@ -231,10 +198,7 @@ export async function summarizeGitChangesWithGeminiFlash(
       usageId: 'git-change-summarizer',
       profileId: 'gemini-flash-summarizer',
     }));
-  let text = '';
-  for await (const chunk of client.streamChat(request)) {
-    if (chunk.type === 'text') text += chunk.text;
-  }
+  const text = await collectStreamedText(client, request);
 
   const redacted = maskSecrets(text, options.secrets).trim();
   if (!redacted) return undefined;
