+++ packages/agent-sdk-ts/src/sdk/runtime/terminalObservationSummarizer.ts
@@ -1,6 +1,13 @@
 import type { ChatCompletionRequest, LLMClient } from '../llm';
 import { getGeminiClient } from './geminiClient';
 import type { SecretRegistry } from './SecretRegistry';
+import {
+  clipTextMiddle,
+  collectStreamedText,
+  maskSecrets,
+  resolveNonNegativeIntOption,
+  truncateSummary,
+} from './summarizerCommon';
 
 export interface TerminalObservationInput {
   command: string;
@@ -25,50 +32,12 @@ const DEFAULT_MAX_PROMPT_CHARS = 6_000;
 const DEFAULT_MAX_SUMMARY_CHARS = 5_000;
 const CLIP_MARKER = '<output clipped>';
 
-const resolveNonNegativeIntOption = (value: unknown, defaultValue: number): number => {
-  if (value === undefined) return defaultValue;
-  if (typeof value !== 'number' || !Number.isFinite(value)) return defaultValue;
-  return Math.max(0, Math.trunc(value));
-};
-
-const clipTextMiddle = (text: string, maxChars: number): string => {
-  if (maxChars <= 0) return '';
-  if (text.length <= maxChars) return text;
-  const markerBudget = CLIP_MARKER.length + 2;
-  if (maxChars < markerBudget) return text.slice(0, maxChars);
-  const available = maxChars - markerBudget;
-  const headLen = Math.ceil(available / 2);
-  const tailLen = Math.floor(available / 2);
-  const head = text.slice(0, headLen);
-  const tail = tailLen === 0 ? '' : text.slice(-tailLen);
-  return `${head}\n${CLIP_MARKER}\n${tail}`;
-};
-
-const maskSecrets = (text: string, secrets: SecretRegistry): string => {
-  let masked = text;
-  const values = secrets
-    .getRegisteredValues()
-    .map((value) => value.trim())
-    .filter((value) => value.length > 0)
-    .sort((a, b) => b.length - a.length);
-  for (const value of values) {
-    masked = masked.replaceAll(value, '***');
-  }
-  return masked;
-};
-
 const summarizeExitCodeFallback = (exitCode: TerminalObservationInput['exit_code']): string => {
   const normalized =
     typeof exitCode === 'number' ? exitCode.toString() : typeof exitCode === 'string' && exitCode.trim() ? exitCode.trim() : 'unknown';
   return normalized === '0' ? 'Done.' : `Done (exit code ${normalized}).`;
 };
 
-const truncateSummary = (text: string, maxChars: number): string => {
-  if (maxChars <= 0) return '';
-  if (text.length <= maxChars) return text;
-  if (maxChars === 1) return '…';
-  return text.slice(0, maxChars - 1) + '…';
-};
 
 export async function summarizeTerminalObservationWithGeminiFlash(
   input: TerminalObservationInput,
@@ -83,7 +52,7 @@ export async function summarizeTerminalObservationWithGeminiFlash(
   const stderr = typeof input.stderr === 'string' ? input.stderr.trimEnd() : '';
   const output = stdout && stderr ? `${stdout}\n${stderr}` : stdout || stderr;
 
-  const redactedOutput = clipTextMiddle(maskSecrets(output, options.secrets), maxOutputChars);
+  const redactedOutput = clipTextMiddle(maskSecrets(output, options.secrets), maxOutputChars, CLIP_MARKER);
   const prompt = [
     'Write a concise (1–3 sentence) summary of a terminal command execution performed by an autonomous coding agent.',
     '- Focus on the action taken and key outcome.',
@@ -98,7 +67,7 @@ export async function summarizeTerminalObservationWithGeminiFlash(
     redactedOutput || '(empty)',
   ].join('\n');
 
-  const safePrompt = clipTextMiddle(maskSecrets(prompt, options.secrets), maxPromptChars);
+  const safePrompt = clipTextMiddle(maskSecrets(prompt, options.secrets), maxPromptChars, CLIP_MARKER);
   const request: ChatCompletionRequest = {
     systemPrompt: 'You summarize terminal tool results for an IDE UI.',
     messages: [{ role: 'user', content: [{ type: 'text', text: safePrompt }] }],
@@ -111,10 +80,7 @@ export async function summarizeTerminalObservationWithGeminiFlash(
         usageId: 'terminal-observation-summarizer',
         profileId: 'gemini-flash-summarizer',
       }));
-    let text = '';
-    for await (const chunk of client.streamChat(request)) {
-      if (chunk.type === 'text') text += chunk.text;
-    }
+    const text = await collectStreamedText(client, request);
     const summary = maskSecrets(text, options.secrets).trim();
     if (!summary) return fallback;
     return truncateSummary(summary, maxSummaryChars);
